feat: Add Docker build script for Docker Volume Backup Script
This commit adds a new bash script to build a Docker image for the Docker Volume Backup Script. The script prompts the user for a version number and builds an image with that tag. It also displays information about the built image and prompts the user to push it to a registry.